move `ModelStages` to enums
