--- src/zenml/model/model_stages.py
@@ -1,25 +1,0 @@
-#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at:
-#
-#       https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-#  or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-"""ModelStages lists supported stages of a Model Version."""
-
-from zenml.utils.enum_utils import StrEnum
-
-
-class ModelStages(StrEnum):
-    """All possible stages of a Model Version."""
-
-    NONE = "none"
-    STAGING = "staging"
-    PRODUCTION = "production"
-    ARCHIVED = "archived"
