+++ src/zenml/models/__init__.py
@@ -165,9 +165,6 @@ from zenml.models.model_models import (
     ModelVersionUpdateModel,
 )
 
-from zenml.models.model_base_model import ModelConfigModel
-
-from zenml.model.model_stages import ModelStages
 
 ComponentResponseModel.update_forward_refs(
     UserResponseModel=UserResponseModel,
@@ -313,9 +310,6 @@ ModelVersionPipelineRunResponseModel.update_forward_refs(
     UserResponseModel=UserResponseModel,
     WorkspaceResponseModel=WorkspaceResponseModel,
 )
-
-ModelVersionUpdateModel.update_forward_refs(ModelStages=ModelStages)
-ModelConfigModel.update_forward_refs(ModelStages=ModelStages)
 
 __all__ = [
     "ArtifactRequestModel",
