+++ src/zenml/models/model_base_model.py
@@ -13,14 +13,12 @@
 #  permissions and limitations under the License.
 """Model base model to support Model WatchTower feature."""
 
-from typing import TYPE_CHECKING, List, Optional, Union
+from typing import List, Optional, Union
 
 from pydantic import BaseModel, Field
 
+from zenml.enums import ModelStages
 from zenml.models.constants import STR_FIELD_MAX_LENGTH, TEXT_FIELD_MAX_LENGTH
-
-if TYPE_CHECKING:
-    from zenml.model.model_stages import ModelStages
 
 
 class ModelBaseModel(BaseModel):
@@ -73,7 +71,7 @@ class ModelConfigModel(ModelBaseModel):
     recovery: Whether to keep failed runs with new versions for later recovery from it.
     """
 
-    version: Optional[Union["ModelStages", str]] = Field(
+    version: Optional[Union[ModelStages, str]] = Field(
         default=None,
         description="Model version or stage is optional and points model context to a specific version/stage, "
         "if skipped and `create_new_model_version` is False - latest model version will be used.",
