+++ src/zenml/zen_stores/sql_zen_store.py
@@ -62,6 +62,7 @@ from zenml.constants import (
 from zenml.enums import (
     ExecutionStatus,
     LoggingLevels,
+    ModelStages,
     SecretScope,
     SorterOps,
     StackComponentType,
@@ -76,7 +77,6 @@ from zenml.exceptions import (
 )
 from zenml.io import fileio
 from zenml.logger import get_console_handler, get_logger, get_logging_level
-from zenml.model.model_stages import ModelStages
 from zenml.models import (
     ArtifactFilterModel,
     ArtifactRequestModel,
