+++ tests/integration/functional/model/test_model_config.py
@@ -17,7 +17,8 @@ import pytest
 
 from zenml.client import Client
 from zenml.constants import RUNNING_MODEL_VERSION
-from zenml.model import ModelConfig, ModelStages
+from zenml.enums import ModelStages
+from zenml.model import ModelConfig
 from zenml.models import ModelRequestModel, ModelVersionRequestModel
 
 MODEL_NAME = "super_model"
